// src/types/type.ts
export interface Article {
    source: { id: string | null; name: string };
    author: string | null;
    title: string;
    description: string;
    url: string;
    urlToImage: string;
    publishedAt: string;
    content: string;
}
export interface NewsState {
    newsData: Article[];
    loading: boolean;
    error: string | null;
}

export interface Currency {
    [key: string]: number;
}
export interface CurrencyState {
    currencyData: Currency;
    loading: boolean;
    error: string | null;
}